Add files via upload
This mod contains: the Mario Dimension that can be made with red concrete and be lit up with Mario's hat. You can make the Mario hat with 6 paper and 1 nametag. in the crafting table you need to put the six paper up and the name tag at the lowest middle point. The food onion was added. A farmer can trade it with 6 coal for 2 onions and the farmer has to be lever 2. The eye ball food was added that can be made with 8 rotten flesh and 1 spider eye. To make this eye you need to place the 8 rotten flesh around the spider eye and the spider eye is in the middle.